refactor(runner|worker): ChainedTaskRunner & Worker execution
The ChainedTaskRunner now use a - clean - worker to execute tasks
The worker options handling has been improved
