+++ src/Worker/AbstractWorker.php
@@ -1,0 +1,189 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SchedulerBundle\Worker;
+
+use DateTimeImmutable;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use SchedulerBundle\Event\TaskExecutedEvent;
+use SchedulerBundle\Event\TaskExecutingEvent;
+use SchedulerBundle\Event\WorkerRestartedEvent;
+use SchedulerBundle\Exception\LogicException;
+use SchedulerBundle\Runner\RunnerInterface;
+use SchedulerBundle\SchedulerInterface;
+use SchedulerBundle\Task\TaskExecutionTrackerInterface;
+use SchedulerBundle\Task\TaskInterface;
+use SchedulerBundle\Task\TaskList;
+use SchedulerBundle\Task\TaskListInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Contracts\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+use Throwable;
+
+/**
+ * @author Guillaume Loulier <contact@guillaumeloulier.fr>
+ */
+abstract class AbstractWorker implements WorkerInterface
+{
+    /**
+     * @var array<string, bool|int|null|TaskInterface>
+     */
+    protected array $options = [];
+    private ?EventDispatcherInterface $eventDispatcher;
+    private TaskListInterface $failedTasks;
+    private LoggerInterface $logger;
+    private SchedulerInterface $scheduler;
+    private TaskExecutionTrackerInterface $tracker;
+
+    public function __construct(
+        SchedulerInterface $scheduler,
+        TaskExecutionTrackerInterface $tracker,
+        ?EventDispatcherInterface $eventDispatcher,
+        ?LoggerInterface $logger
+    ) {
+        $this->scheduler = $scheduler;
+        $this->tracker = $tracker;
+        $this->eventDispatcher = $eventDispatcher;
+        $this->logger = $logger ?? new NullLogger();
+        $this->failedTasks = new TaskList();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function restart(): void
+    {
+        $this->stop();
+        $this->options['isRunning'] = false;
+        $this->failedTasks = new TaskList();
+        $this->options['shouldStop'] = false;
+
+        $this->dispatch(new WorkerRestartedEvent($this));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function stop(): void
+    {
+        $this->options['shouldStop'] = true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isRunning(): bool
+    {
+        return $this->options['isRunning'];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFailedTasks(): TaskListInterface
+    {
+        return $this->failedTasks;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLastExecutedTask(): ?TaskInterface
+    {
+        return $this->options['lastExecutedTask'];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getOptions(): array
+    {
+        return $this->options;
+    }
+
+    /**
+     * @throws Throwable {@see SchedulerInterface::getDueTasks()}
+     */
+    protected function getTasks(array $tasks): array
+    {
+        $tasks = [] !== $tasks ? $tasks : $this->scheduler->getDueTasks();
+
+        return is_array($tasks) ? $tasks : iterator_to_array($tasks);
+    }
+
+    protected function handleTask(RunnerInterface $runner, TaskInterface $task): void
+    {
+        $this->dispatch(new TaskExecutingEvent($task));
+
+        $task->setArrivalTime(new DateTimeImmutable());
+        $task->setExecutionStartTime(new DateTimeImmutable());
+
+        $this->tracker->startTracking($task);
+        $output = $runner->run($task, $this);
+        $this->tracker->endTracking($task);
+        $task->setExecutionEndTime(new DateTimeImmutable());
+        $task->setLastExecution(new DateTimeImmutable());
+
+        $this->dispatch(new TaskExecutedEvent($task, $output));
+    }
+
+    protected function checkTaskState(TaskInterface $task): bool
+    {
+        if (TaskInterface::UNDEFINED === $task->getState()) {
+            throw new LogicException('The task state must be defined in order to be executed!');
+        }
+
+        if (in_array($task->getState(), [TaskInterface::PAUSED, TaskInterface::DISABLED], true)) {
+            $this->logger->info(sprintf('The following task "%s" is paused|disabled, consider enable it if it should be executed!', $task->getName()), [
+                'name' => $task->getName(),
+                'expression' => $task->getExpression(),
+                'state' => $task->getState(),
+            ]);
+
+            return false;
+        }
+
+        return true;
+    }
+
+    protected function getSleepDuration(): int
+    {
+        $dateTimeImmutable = new DateTimeImmutable('+ 1 minute', $this->scheduler->getTimezone());
+        $updatedNextExecutionDate = $dateTimeImmutable->setTime((int) $dateTimeImmutable->format('H'), (int) $dateTimeImmutable->format('i'));
+
+        return (new DateTimeImmutable('now', $this->scheduler->getTimezone()))->diff($updatedNextExecutionDate)->s + $this->options['sleepDurationDelay'];
+    }
+
+    protected function dispatch(Event $event): void
+    {
+        if (!$this->eventDispatcher instanceof EventDispatcherInterface) {
+            return;
+        }
+
+        $this->eventDispatcher->dispatch($event);
+    }
+
+    protected function configure(array $options): void
+    {
+        $optionsResolver = new OptionsResolver();
+        $optionsResolver->setDefaults([
+            'executedTasksCount' => 0,
+            'isRunning' => false,
+            'lastExecutedTask' => null,
+            'sleepDurationDelay' => 1,
+            'sleepUntilNextMinute' => false,
+            'shouldStop' => false,
+        ]);
+
+        $optionsResolver->setAllowedTypes('executedTasksCount', 'int');
+        $optionsResolver->setAllowedTypes('isRunning', 'bool');
+        $optionsResolver->setAllowedTypes('lastExecutedTask', [TaskInterface::class, 'null']);
+        $optionsResolver->setAllowedTypes('sleepDurationDelay', 'int');
+        $optionsResolver->setAllowedTypes('sleepUntilNextMinute', 'bool');
+        $optionsResolver->setAllowedTypes('shouldStop', 'bool');
+
+        $this->options = $optionsResolver->resolve($options);
+    }
+}
